comment model moved to its app
